# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafecitos API WEB",
        "description": "API for Cafecitos",
        "contact": {
            "email": "devcitobo@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "operationId": "6d8ae74dcafe19c4813904d4ea9e4c57",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "9f0b814cd8d69075ed70d4cc79c9f417",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "description",
                                    "client_price",
                                    "waiter_price",
                                    "animation_price",
                                    "waiter_commission",
                                    "is_active"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "client_price": {
                                        "type": "number"
                                    },
                                    "waiter_price": {
                                        "type": "number"
                                    },
                                    "animation_price": {
                                        "type": "number"
                                    },
                                    "waiter_commission": {
                                        "type": "number"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a product by ID",
                "operationId": "f6410f1810b0a90ba3d72d5b8250a27d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "14b6ff17cc6ce7298a06271a8bc89539",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "description",
                                    "client_price",
                                    "waiter_price",
                                    "animation_price",
                                    "waiter_commission",
                                    "is_active"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "client_price": {
                                        "type": "number"
                                    },
                                    "waiter_price": {
                                        "type": "number"
                                    },
                                    "animation_price": {
                                        "type": "number"
                                    },
                                    "waiter_commission": {
                                        "type": "number"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/products/{id}/status": {
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Change product status",
                "operationId": "361dac520776500b857c6eb7b61b90bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_active"
                                ],
                                "properties": {
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/products/{id}/image": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a product image",
                "operationId": "b7593996ea5c8cc128b04d4639b9cb0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image file",
                        "content": {
                            "image/png": {}
                        }
                    },
                    "404": {
                        "description": "Image not found"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Upload a product image",
                "operationId": "42f4d7be62f135d819822f1cbb0f769c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image uploaded"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
        "/api/products/types": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all product types",
                "operationId": "5f17be47cd2b4256fa52d80dc69bbaa8",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "csrf": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "csrf": {
                "type": "apiKey",
                "description": "Para obtener este token, primero asegúrate de haber iniciado sesión. Luego, ejecuta el endpoint GET /sanctum/csrf-cookie. Finalmente, abre las herramientas de desarrollador de tu navegador, busca la cookie 'XSRF-TOKEN', copia su valor y pégalo aquí. Este token es necesario para todas las peticiones que modifican datos (POST, PUT, PATCH, DELETE).",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
...
}